Modified Readme and Added UTF tests
For this branch I modified the readme to be a demo script for the LabVIEW 2018 CLI. I included instructions for running 6 different CLI commands using the VIs and file paths currently saved on the demo laptop. The only command I could not get to run successfully was the VI Analyzer command. The error I ran into is documented in the readme